+++ sewer/tests/test_auth.py
@@ -46,7 +46,8 @@ class TestAuth01(unittest.TestCase):
     def test07_accepts_log_level(self):
         self.assertTrue(pbj(LOG_LEVEL="INFO"))
 
-    @unittest.skip("current implementation allows both")
+    # current implementation allows both parameters :-(
+    @unittest.expectedFailure
     def test08_rejects_logger_and_log_level(self):
         with self.assertRaises(ValueError):
             pbj(logger=lib.create_logger("", "INFO"), LOG_LEVEL="INFO")
@@ -115,14 +116,14 @@ class TestAuthDNS(unittest.TestCase):
 
     def test04_without_alias(self):
         p = auth.DNSProviderBase()
-        chal = {"domain": "example.com"}
+        chal = {"ident_value": "example.com"}
         self.assertTrue(
             p.cname_domain(chal) is None and p.target_domain(chal) == "_acme-challenge.example.com"
         )
 
     def test05_with_alias(self):
         p = auth.DNSProviderBase(alias="valid.com")
-        chal = {"domain": "example.com"}
+        chal = {"ident_value": "example.com"}
         self.assertTrue(
             p.target_domain(chal) == "example.com.valid.com"
             and p.cname_domain(chal) == "_acme-challenge.example.com"